refactor(core): move runGeneration onto createKubb as Kubb.generate
Fold the free runGeneration() function into a generate() method on the Kubb
instance returned by createKubb, so a host calls createKubb(config, { hooks })
.generate(...) instead of threading the resolved config and hooks through a
second entry point. The method runs on this.config, the config createKubb has
already resolved (root default included), which fixes the unplugin path where
the raw config left root undefined.

The CLI, the bundler plugin, and the MCP tool now call generate(); the free
function and its test are removed. Generated output is unchanged.
