+++ packages/core/src/index.ts
@@ -2,8 +2,7 @@ export { createAdapter } from './createAdapter.ts'
 export { applyConfigDefaults } from './applyConfigDefaults.ts'
 export { Diagnostics } from './Diagnostics.ts'
 export { createKubb } from './createKubb.ts'
-export { runGeneration } from './runGeneration.ts'
-export type { GenerationPhase, GenerationResult, RunGenerationOptions } from './runGeneration.ts'
+export type { GenerateOptions, GenerationPhase, GenerationResult } from './createKubb.ts'
 export { createReporter, logLevel } from './createReporter.ts'
 export { cliReporter } from './reporters/cliReporter.ts'
 export { fileReporter } from './reporters/fileReporter.ts'
